split Topics.cs into TopicInfo.cs & TopicsController.cs
