+++ components/Topics/TopicsController.cs
@@ -30,107 +30,6 @@ using System.Text.RegularExpressions;
 
 namespace DotNetNuke.Modules.ActiveForums
 {
-#region TopicInfo Class
-	public class TopicInfo
-	{
-		public TopicInfo()
-		{
-			Content = new Content();
-			Security = new PermissionInfo();
-			Author = new Author();
-		}
-
-        #region Private Members
-        #endregion
-        #region Public Properties
-        public int TopicId { get; set; }
-        public int ContentId { get; set; }
-        public int ViewCount { get; set; }
-        public int ReplyCount { get; set; }
-        public bool IsLocked { get; set; }
-        public bool IsPinned { get; set; }
-        public string TopicIcon { get; set; }
-        public int StatusId { get; set; }
-        public bool IsApproved { get; set; }
-        public bool IsDeleted { get; set; }
-        public bool IsAnnounce { get; set; }
-        public bool IsArchived { get; set; }
-        public TopicTypes TopicType { get; set; }
-        public DateTime AnnounceStart { get; set; }
-        public DateTime AnnounceEnd { get; set; }
-        public Content Content { get; set; }
-        public PermissionInfo Security { get; set; }
-        public Author Author { get; set; }
-        public string Tags { get; set; }
-        public int Priority { get; set; } = 0;
-        public string Categories { get; set; } = string.Empty;
-        public string TopicUrl { get; set; } = string.Empty;
-        public string ForumURL { get; set; } = string.Empty;
-        public string TopicData { get; set; } = string.Empty;
-
-        public string URL
-		{
-			get
-			{
-				if (! (string.IsNullOrEmpty(TopicUrl)) && ! (string.IsNullOrEmpty(ForumURL)))
-				{
-					if (TopicUrl.StartsWith("/") & ForumURL.EndsWith("/"))
-					{
-						ForumURL = ForumURL.Substring(0, ForumURL.Length - 1);
-						if (! (ForumURL.StartsWith("/")))
-						{
-							ForumURL = "/" + ForumURL;
-						}
-					}
-					return ForumURL + TopicUrl;
-				}
-				else
-				{
-					return string.Empty;
-				}
-			}
-		}
-		public List<PropertiesInfo> TopicProperties
-		{
-			get
-			{
-				if (TopicData == string.Empty)
-				{
-					return null;
-				}
-				else
-				{
-					List<PropertiesInfo> pl = new List<PropertiesInfo>();
-					XmlDocument xDoc = new XmlDocument();
-					xDoc.LoadXml(TopicData);
-					if (xDoc != null)
-					{
-						System.Xml.XmlNode xRoot = xDoc.DocumentElement;
-						System.Xml.XmlNodeList xNodeList = xRoot.SelectNodes("//properties/property");
-						if (xNodeList.Count > 0)
-						{
-							int i = 0;
-							for (i = 0; i < xNodeList.Count; i++)
-							{
-								string pName = Utilities.HTMLDecode(xNodeList[i].ChildNodes[0].InnerText);
-								string pValue = Utilities.HTMLDecode(xNodeList[i].ChildNodes[1].InnerText);
-								int pId = Convert.ToInt32(xNodeList[i].Attributes["id"].Value);
-								PropertiesInfo p = new PropertiesInfo();
-								p.Name = pName;
-								p.DefaultValue = pValue;
-								p.PropertyId = pId;
-								pl.Add(p);
-							}
-						}
-					}
-					return pl;
-				}
-			}
-		}
-
-#endregion
-	}
-#endregion
 #region Topics Controller
 	public class TopicsController : DotNetNuke.Entities.Modules.ISearchable
 	{
